Use response time for probability weights.
The three most recent answers for a word are used to calculate the probability
weight of that word being shown again, with the more recent answers weighted
more. Correct answers, adjusted for kana character length, are categorized into
slow, medium, and fast responses. The current streak scoring is replaced with a
speed based scoring shown at the top of the screen. Similar to the previous
streak scoring, displaying scores is optional. Emojis representing fast
(rabbit), medium (pig), slow (turtle), and incorrect (snake) are included in the
response status box. An optional stats button will display a list of words seen
so far with their conjugation and score emojis.

Other changes:

 - Decreased probability weights within iku and suru "word groups" is no longer used.

 - Uncomment the calls to the new logResponseTimeDetails and
   logProbabilityWeights functions in the main.js file in order to see these
   details in the console log.

 - The .devcontainer/devcontainer.json and .vscode/launch.json files are
   included to allow node and npm installations to be isolated to a Docker dev
   container while supporting hot reload and debugging support.

 - The Dockerfile and associated .dockerignore files are included if deployment
   uses a Docker container.